Update create reviewer function

+++ pages/admin/add-reviewer.tsx
@@ -4,9 +4,55 @@ import * as XLSX from 'xlsx'
 import AdminLayout from '../../layouts/admin/admin-layout'
 import { createReviewer } from '../api/createReviewer'
 import firebase from '../../firebase'
+import { Reviewer } from '../../classes/reviewer'
 
 export default function AddReviewer() {
   const [file, setFile] = useState<File>()
+  const [addedReviewers, setAddedReviewers] = useState<Reviewer[]>([])
+  const [removedReviewers, setRemovedReviewers] = useState<Reviewer[]>([])
+
+  const createAccount = async (secondaryFirebaseApp, reviewer, password) => {
+    await secondaryFirebaseApp
+      .auth()
+      .createUserWithEmailAndPassword(reviewer[0], password)
+      .then(async (result: firebase.auth.UserCredential) => {
+        secondaryFirebaseApp.auth().signOut()
+
+        // add reviewer data to firestore database
+        createReviewer(
+          result.user.uid,
+          reviewer[0],
+          reviewer[1],
+          reviewer[2],
+          reviewer[3],
+        )
+
+        setAddedReviewers((prevAddedReviewers) => [
+          ...prevAddedReviewers,
+          new Reviewer(reviewer[0], reviewer[2], reviewer[3], reviewer[1]),
+        ])
+
+        let templateParams = {
+          email: reviewer[0],
+          personal_email: reviewer[1],
+          name: reviewer[2],
+          set: reviewer[3],
+          password: password,
+        }
+        emailjs.send(
+          process.env.EMAILJS_SERVICE_ID,
+          process.env.EMAILJS_REVIEWER_TEMPLATE_ID,
+          templateParams,
+          process.env.EMAILJS_USER_ID,
+        )
+      })
+      .catch(() => {
+        setRemovedReviewers((prevRemovedReviewers) => [
+          ...prevRemovedReviewers,
+          new Reviewer(reviewer[0], reviewer[2], reviewer[3], reviewer[1]),
+        ])
+      })
+  }
 
   const proceed = () => {
     const reader = new FileReader()
@@ -33,7 +79,15 @@ export default function AddReviewer() {
       // Get all reviewer details as array from worksheet
       let reviewers = XLSX.utils.sheet_to_json(sheet, { header: 1 })
 
-      reviewers.forEach((reviewer) => {
+      reviewers.forEach(async (reviewer, index) => {
+        if (!index) return // leave first row
+
+        // reviewer details
+        // reviewer[0] represents email from which account is created
+        // reviewer[1] represents personal email on which details are to be mailed
+        // reviewer[2] represents name of reviewer
+        // reviewer[3] represents reviewer set
+
         // Generate a random password
         let charset =
           'abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ0123456789'
@@ -42,39 +96,14 @@ export default function AddReviewer() {
           password += charset.charAt(Math.floor(Math.random() * 62))
 
         // Create account of reviewer using signup on new firebase app
-        secondaryFirebaseApp
-          .auth()
-          .createUserWithEmailAndPassword(reviewer[0], password)
-          .then(async (result: firebase.auth.UserCredential) => {
-            secondaryFirebaseApp.auth().signOut()
+        await createAccount(secondaryFirebaseApp, reviewer, password)
 
-            // add reviewer data to firestore database
-            createReviewer(
-              result.user.uid,
-              reviewer[0],
-              reviewer[1],
-              reviewer[2],
-            )
-
-            let templateParams = {
-              name: reviewer[1],
-              email: reviewer[0],
-              password: password,
-            }
-            emailjs.send(
-              process.env.EMAILJS_SERVICE_ID,
-              process.env.EMAILJS_REVIEWER_TEMPLATE_ID,
-              templateParams,
-              process.env.EMAILJS_USER_ID,
-            )
-          })
-          .catch((e) => console.log(e))
+        if (index == reviewers.length - 1) {
+          alert('Added all reviewers')
+          // Delete firebase application created above
+          secondaryFirebaseApp.delete()
+        }
       })
-
-      // Delete firebase application created above
-      setTimeout(() => {
-        secondaryFirebaseApp.delete()
-      }, 5000)
     }
 
     reader.readAsBinaryString(file)
@@ -96,13 +125,73 @@ export default function AddReviewer() {
           <input
             type="submit"
             onClick={proceed}
-            className="text-white text-base md:text-lg bg-blue-850 ml-2 px-2 rounded-lg flex flex-row items-center"
+            className="text-white text-base md:text-lg bg-blue-850 ml-2 px-2 rounded-lg flex flex-row items-center cursor-pointer"
           />
         </div>
         <p className="text-red-850 text-center mb-40">
           Note: Provide the excel file containing the data in the specified
           format.
         </p>
+
+        {!addedReviewers.length ? null : (
+          <div className="my-10">
+            <h1 className="text-sm sm:text-xl md:text-2xl bg-blue-850 my-10 text-white text-center font-extrabold py-2 rounded-tl-3xl rounded-br-3xl">
+              Added Reviewers
+            </h1>
+            <div className="flex justify-center">
+              <table className="border p-2">
+                <thead>
+                  <tr>
+                    <th className="border p-2">Name</th>
+                    <th className="border p-2">Email</th>
+                    <th className="border p-2">Set</th>
+                    <th className="border p-2">Personal Email</th>
+                  </tr>
+                </thead>
+                <tbody>
+                  {addedReviewers.map((reviewer: Reviewer) => (
+                    <tr>
+                      <td className="border p-2">{reviewer.name}</td>
+                      <td className="border p-2">{reviewer.email}</td>
+                      <td className="border p-2">{reviewer.set}</td>
+                      <td className="border p-2">{reviewer.personal_email}</td>
+                    </tr>
+                  ))}
+                </tbody>
+              </table>
+            </div>
+          </div>
+        )}
+
+        {!removedReviewers.length ? null : (
+          <div className="my-10">
+            <h1 className="text-sm sm:text-xl md:text-2xl bg-blue-850 my-10 text-white text-center font-extrabold py-2 rounded-tl-3xl rounded-br-3xl">
+              Removed Reviewers
+            </h1>
+            <div className="flex justify-center">
+              <table className="border p-2">
+                <thead>
+                  <tr>
+                    <th className="border p-2">Name</th>
+                    <th className="border p-2">Email</th>
+                    <th className="border p-2">Set</th>
+                    <th className="border p-2">Personal Email</th>
+                  </tr>
+                </thead>
+                <tbody>
+                  {removedReviewers.map((reviewer: Reviewer) => (
+                    <tr>
+                      <td className="border p-2">{reviewer.name}</td>
+                      <td className="border p-2">{reviewer.email}</td>
+                      <td className="border p-2">{reviewer.set}</td>
+                      <td className="border p-2">{reviewer.personal_email}</td>
+                    </tr>
+                  ))}
+                </tbody>
+              </table>
+            </div>
+          </div>
+        )}
       </div>
     </AdminLayout>
   )
